Jun 25, 2024, 10:17 PM

+++ main.py
@@ -16,9 +16,9 @@ while not validanswer:
         validanswer = True
 validanswer = False
 while not validanswer:
-    wall2width = float(input("Enter width of wall 2 (m): "))
+    wall2width = input("Enter width of wall 2 (m): ")
     if wall2width == type(float) or type(int):
-        wall2width = int(wall2width)
+        wall2width = float(wall2width)
         if float(wall2width) > 0:
             validanswer = True
     else:
